Factor out a table.js from explorer.js.
